+++ tests/conftest.py
@@ -49,16 +49,16 @@ DIFF_TEST_INPUTS = {
 example_ccli = """Build My LifeSongSelect logo
 Brett Younker | Karl Martin | Kirby Kaple | Matt Redman | Pat Barrett
 (based on the recording by Pat Barrett)
-Key - G | Tempo - 70 | Time - 4/4 
+Key - G | Tempo - 70 | Time - 4/4
 
 Intro
-|  G          |  C2         |  G/B         |  C2         |  
+|  G          |  C2         |  G/B         |  C2         |
 
 Verse 1
 G     Worthy of ev'ry C2song we could ever sing
 G/B     Worthy of all the C2praise we could ever bring
 G     Worthy of ev'ry C2breath we could ever breathe
-We live for G/BYou         C2 
+We live for G/BYou         C2
 
 Verse 2
 G     Jesus the name a - C2bove ev'ry other name
@@ -68,15 +68,15 @@ We live for G/BYou,    we live for C2You
 
 Chorus
 C2Holy there is no one Amlike You
-There is none be - G/Dside You open up my Em7eyes in wonder 
-And C2show me who You are and Amfill me 
+There is none be - G/Dside You open up my Em7eyes in wonder
+And C2show me who You are and Amfill me
 With Your heart and G/Dlead me in Your love to Em7those around me
 
 Bridge
-CI will build my Dlife upon Your Em7love 
+CI will build my Dlife upon Your Em7love
 It is a G/Bfirm foundation
-CI will put my Dtrust in You a - Em7lone 
-And I will G/Bnot be shaken  (C)   (Last time) 
+CI will put my Dtrust in You a - Em7lone
+And I will G/Bnot be shaken  (C)   (Last time)
 
 CCLI Song # 7070345
 
@@ -122,6 +122,7 @@ def tmp_csv_dicts() -> list[list[dict[str, int]]]:
             data.append(d)
         tmp_dict_list.append(data)
     return tmp_dict_list
+
 
 @pytest.fixture
 def tmp_ccli_files(tmp_path: Path) -> list[Path]: