+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Updated
+
+- updated dependency versions
+
 ## [0.7.0] - 2026-02-22
 
 ### Changed