+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Fixed
+
+- correctly added new script ccli2chpro to pyproject.toml
+
 ## [0.8.0] - 2026-06-23
 
 ### Added